First working version
This version does everything that the tuscany::cpp extension does,
but uses the SelfProtrait reflection library instead of relying
on specially generated stubs.

I've only tested it with the calculato sample, but the composition
of the CalculatorComponent and the DivideComponent version is already
working.

The next step is to make the dependency injection work.